nlw-spacetime - Web and mobile authentication

+++ NLW/2023/SpaceTime/server/src/routes/memories.ts
@@ -3,10 +3,17 @@ import { prisma } from "../lib/prisma";
 import { z } from 'zod'
 
 export async function memoriesRoute(app: FastifyInstance) {
-    app.get('/memories', async (request) => {
+
+    app.addHook('preHandler', async (request) => {
         await request.jwtVerify()
+    })
+
+    app.get('/memories', async (request) => {
 
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: request.user.sub,
+            },
             orderBy: {
                 createdAt: 'asc',
             },
@@ -21,7 +28,7 @@ export async function memoriesRoute(app: FastifyInstance) {
         })
     })
 
-    app.get('/memories/:id', async (request) => {
+    app.get('/memories/:id', async (request, reply) => {
         // const { id } = request.params
         const paramsSchema = z.object({
             id: z.string().uuid(),
@@ -34,6 +41,10 @@ export async function memoriesRoute(app: FastifyInstance) {
                 id,
             },
         })
+
+        if (!memory.isPublic && memory.userId != request.user.sub) {
+            return reply.status(401).send()
+        }
 
         return memory
     })
@@ -52,7 +63,7 @@ export async function memoriesRoute(app: FastifyInstance) {
                 content,
                 coverUrl,
                 isPublic,
-                userId: 'b1574248-1f95-4078-be20-3c21bae96d13'
+                userId: request.user.sub
             },
         })
 
@@ -60,7 +71,7 @@ export async function memoriesRoute(app: FastifyInstance) {
         
     })
 
-    app.put('/memories/:id', async (request) => {
+    app.put('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
@@ -74,7 +85,17 @@ export async function memoriesRoute(app: FastifyInstance) {
 
         const {content, coverUrl, isPublic} = bodySchema.parse(request.body)
 
-        const memory =await prisma.memory.update({
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        if (memory.userId != request.user.sub) {
+            return reply.status(401).send()
+        }
+
+        await prisma.memory.update({
             where: {
                 id,
             },
@@ -89,12 +110,22 @@ export async function memoriesRoute(app: FastifyInstance) {
         return memory
     })
 
-    app.delete('/memories/:id', async (request) => {
+    app.delete('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
 
         const {id} = paramsSchema.parse(request.params)
+
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        if (memory.userId != request.user.sub) {
+            return reply.status(401).send()
+        }
 
         await prisma.memory.delete({
             where: {
